Fix QCC compiler warnings in libjpeg-turbo

+++ 3dparty/libjpeg-turbo/jdphuff.c
@@ -205,7 +205,7 @@ start_pass_phuff_decoder (j_decompress_ptr cinfo)
 #ifdef AVOID_TABLES
 
 #define NEG_1 ((unsigned)-1)
-#define HUFF_EXTEND(x,s)  ((x) < (1<<((s)-1)) ? (x) + (((NEG_1)<<(s)) + 1) : (x))
+#define HUFF_EXTEND(x,s)  ((x) < ((int) (1<<((s)-1))) ? (int) ((x) + (((NEG_1)<<(s)) + 1)) : (x))
 
 #else
